Massive changes to UI of the project, plus more
We take out materializecss as the primary front end framework and use the PrimeNG library. Create the shared module to import the PrimeNG modules when needed, also create form validation for the register page and the login page. Begin creating http service for the login page as well by creating an authentication module.