Changes/fixes to the login and register components
Removing form validation on the login form, and then implementing the text-mask module for the phone number field inside the register component. Form validation should be complete. Now moving on to create the Http service to test to see if we can send our form from the front end to the backend.