Connect login component to redirect to either the lab or ehos dashboard components
The login and register components are fully functional, so we have the ability to either post a user in JSON format to a db and/or authenticate a user's credentials that are also in JSON format. The login component has a web token to enable data within the application to be readily accessable. Now that routes to the two dashboards from the login are connected, now it is time to work on the various forms for the EHOS/Admin. Next will be an admin section to view and approve registrations.