Changes for sprint
Massive updates. First is the shift to bootstrap 4.0.0 and styling this downloaded package as a flat design with purple being the primary color. Next is a change of the login page. It is now our project logo with updated syntax for bootstrap. Lastly is a fully function ehos approval component that can approve a user to the system. Last thing to achieve is ability to reject spy accounts.